requirements_linux.txt
spidev
st7735
RPi.GPIO
